Add files via upload
This is the  seminar report on the application of PCA in the wine dataset. PCA is a dimensionality reduction technique that can be used to simplify and visualize high-dimensional data. In my report, First i discussed the concepts of mathematics behind it, how to reduce the dimension mathematically, then showed how PCA can be used to identify the most important features in the wine dataset and to cluster the data into different types of wine. I applied PCA to the wine dataset and reduced the dimensionality from 13 to 2 dimensions. I then visualized the data in 2D and identified the clusters of wine. The results showed that the three types of wine (red, yellow and green) were well-separated in the 2D space. This suggests that PCA can be used to cluster wine based on its chemical composition.